* Merged first part of V7

+++ libsrc/Wi/search_in.c
@@ -82,7 +82,7 @@ if (DVC_GREATER == l1) goto gt; \
     iri_id_t i1, i2; \
     dv2 = (db_buf_t) itc->itc_search_params[nth_param];		\
   cl  = key->key_part_cls[nth_key]; \
-    ROW_INT_COL (buf, row, rv, *cl, (iri_id_t) LONG_REF, i1); \
+  ROW_INT_COL (buf, row, rv, *cl, (iri_id_t)(uint32) LONG_REF, i1);	\
   i2 = unbox_iri_id (((caddr_t) dv2));				\
   if (i1 < i2) goto lt; \
   if (i1 > i2) goto gt; \
